requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography

# Data Validation
pydantic==2.5.3

# Configuration
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# Data Processing
pandas
numpy
python-dateutil

# CORS Support (FastAPI内蔵だが明示的に記載)
# fastapi.middleware.cors を使用

# Production Server (Azure App Service用)
# gunicorn with uvicorn worker